+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - PLC4X bindings
 - Shortcuts to access all Controls devices in a magix network.
 - `DeviceClient` properly evaluates lifecycle and logs
+- `PeerConnection` API for direct device-device binary sharing
 
 ### Changed
 - Constructor properties return `DeviceState` in order to be able to subscribe to them